refactor(api): modularizar rutas y modelos de datos de alertas
